کمی مرتب کردن برنامه
