+++ divar.ir/README.md
@@ -1,6 +1,9 @@
 # استخراج آگهی‌های دیوار
 توی این پروژه آگهی‌های سایت دیوار توی یک سرچ خاص و شرایط خاص که خود کاربر به ما داده رو دائم رصد می‌کنیم و اگر تغییری اتفاق بیوفته توی بله و تلگرام به کاربر اطلاع می‌دیم.
 
+## شیوه استفاده
+کافیه فایل ```main.py``` اجرا بشه.
+
 ## کمک به ارتقاء پروژه
 ### موقعیت اول
 اگر بشه کشف کنیم که مقدار عدد جلوی last-post-date چه ارتباطی با تاریخ و زمان به وقت محلی دار خیلی کمک کننده خواهد بود.
