+++ divar.ir/main.py
@@ -1,0 +1,3 @@
+from scrape_divar_page import scrape
+
+scrape()
